backend/requirements.txt
# Django==4.2.9
# djangorestframework==3.14.0
# django-cors-headers==4.3.1
# pymongo==4.6.1
# # djongo==1.3.6
# djongo==1.3.6
# sqlparse==0.2.4
# gunicorn==21.2.0
# python-decouple==3.8
# PyJWT==2.8.0


# Django==4.2.9
# djangorestframework==3.14.0
# django-cors-headers==4.3.1
# pymongo==4.6.1
# # Remove the djongo line
# gunicorn==21.2.0
# python-decouple==3.8
# PyJWT==2.8.0





Django==4.2.9
djangorestframework==3.14.0
django-cors-headers==4.3.1
pymongo==4.6.1
gunicorn==21.2.0
python-decouple==3.8
PyJWT==2.8.0